remove rubric commenting, change multiindex, add parameter names to docstrings
Changed the multiindex to be YEAR, TEAM, PLAYER instead of YEAR, PLAYER, TEAM and then sorted it.
This caused the indexing in the rest of the code to need to be changed.
This added some complication; however, it works now and when the DataFrame is exported to excel the Teams are a joined index.

Also added parameter names to all of the doc strings because I feel like its better.